% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_differences.R
\name{merge_differences}
\alias{merge_differences}
\title{Merge differences across legs}
\usage{
merge_differences(l)
}
\arguments{
\item{l}{list of differences in each leg}
}
\description{
Merge differences across legs
}
