% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst_clustering_threshold.R
\name{mst_clustering_threshold}
\alias{mst_clustering_threshold}
\title{MST Clustering}
\usage{
mst_clustering_threshold(
  corr_matrix,
  leg_names,
  connections,
  corr_threshold = 0.5
)
}
\arguments{
\item{corr_matrix}{correlation matrix for all legs}

\item{leg_names}{character vector of names of legs}

\item{connections}{character vector of leg names which are valid connections}

\item{corr_threshold}{correlation threshold used to split clusters. Default 0.5.}
}
\description{
MST Clustering
}
